feat(mb): added missing e2e Test for UJ-A-003
