+++ frontend/tests/e2e/steps/flight-preparation.steps.ts
@@ -1,7 +1,7 @@
 import { expect } from '@playwright/test'
 import { createBdd } from 'playwright-bdd'
 
-const { Given, When, Then } = createBdd()
+const { When, Then } = createBdd()
 
 // ─── Helpers ───────────────────────────────────────────────────────────────
 
@@ -13,23 +13,6 @@ async function fillStation(page: import('@playwright/test').Page, label: string,
   await input.fill(value)
   await input.dispatchEvent('input')
 }
-
-// ─── Background ────────────────────────────────────────────────────────────
-
-Given('the pilot starts a new flight preparation', async ({ page }) => {
-  await page.goto('/mass-balance')
-  await expect(page.getByLabel('Select aircraft')).toBeVisible()
-})
-
-// ─── When steps — Aircraft selection ───────────────────────────────────────
-
-When('the pilot selects aircraft {string}', async ({ page }, registration: string) => {
-  const select = page.getByLabel('Select aircraft')
-  const option = select.locator('option', { hasText: registration })
-  const label = await option.textContent()
-  await select.selectOption({ label: label!.trim() })
-  await expect(page.locator('.aircraft-label')).toContainText(registration)
-})
 
 // ─── When steps — Happy path (UJ-B-005) ────────────────────────────────────
 
@@ -119,12 +102,3 @@ Then('a critical notification {string} is displayed', async ({ page }, message: 
   await expect(page.locator('.notification--critical', { hasText: message })).toBeVisible()
 })
 
-Then('no critical notifications are displayed', async ({ page }) => {
-  // .notification--critical: semantic severity class; no ARIA role distinguishes critical-level alerts
-  await expect(page.locator('.notification--critical')).toHaveCount(0)
-})
-
-Then('the total mass is within MTOM', async ({ page }) => {
-  // .notification--critical: semantic severity class; no ARIA role distinguishes critical-level alerts
-  await expect(page.locator('.notification--critical', { hasText: 'MTOM Exceeded' })).toHaveCount(0)
-})
